chore: change imagePullPolicy to IfNotPresent
Deployments in a Kind cluster will fail when the image is loaded into the Kind
cluster and the image pull policy is set to Always because the cluster will try
to pull the latest version of the image. It's strongly recommended that deployments
of the components use tagged versions so that deployments won't be automatically
changes as Pods move to new nodes and pull different container versions. Since we
use version tags in most situations, an IfNotPresent image pull policy is a sane default.